Split buttons logic into seperate components
